move slot check tp SPV

+++ contracts/tests/TestMakerDeposit.sol
@@ -346,7 +346,7 @@ contract testMakerDeposit is IORMakerDeposit, VersionAndEnableTime {
 
         require(_challenges[challengeId].statement[msg.sender].challengeTime == 0, "CT");
 
-        require(_challenges[challengeId].result.lastChallengeBlockNum < block.number, "LBN");
+        require(_challenges[challengeId].result.lastChallengeBlockNum < uint64(block.number), "LBN");
         _challenges[challengeId].result.lastChallengeBlockNum = uint64(block.number);
 
         (ruleKey);
@@ -573,12 +573,6 @@ contract testMakerDeposit is IORMakerDeposit, VersionAndEnableTime {
             require(slot == publicInputData.manage_source_chain_info_slot, "CIS");
         }
 
-        // {
-        //     // check sourceChain mainnet token slot
-        //     uint slot = uint(abi.encode(abi.encode(publicInputData.chain_id, publicInputData.token).hash(), 3).hash()) +
-        //         1;
-        //     require(slot == publicInputData.manage_source_chain_mainnet_token_info_slot, "MTS");
-        // }
         statement_s.sourceTxFrom = publicInputData.from;
 
         statement_s.challengeUserRatio = publicInputData.manage_current_challenge_user_ratio;
